+++ jobbergate-agent/jobbergate_agent/jobbergate/update.py
@@ -1,14 +1,26 @@
+import asyncio
 import json
+from itertools import chain
 from typing import List
 
+import msgpack
 from jobbergate_core.tools.sbatch import InfoHandler
 from loguru import logger
 
 from jobbergate_agent.clients.cluster_api import backend_client as jobbergate_api_client
-from jobbergate_agent.jobbergate.schemas import ActiveJobSubmission, SlurmJobData
+from jobbergate_agent.clients.influx import influxdb_client
+from jobbergate_agent.jobbergate.schemas import (
+    ActiveJobSubmission,
+    SlurmJobData,
+    JobSubmissionMetricsMaxResponse,
+    InfluxDBMeasurementDict,
+    InfluxDBPointDict,
+)
 from jobbergate_agent.settings import SETTINGS
 from jobbergate_agent.utils.exception import JobbergateApiError, SbatchError
 from jobbergate_agent.utils.logging import log_error
+from jobbergate_agent.jobbergate.constants import INFLUXDB_MEASUREMENT
+from jobbergate_agent.utils.compute import aggregate_influx_measures
 
 
 async def fetch_job_data(slurm_job_id: int, info_handler: InfoHandler) -> SlurmJobData:
@@ -70,6 +82,89 @@ async def update_job_data(
         response.raise_for_status()
 
 
+async def fetch_influx_data(
+    time: int, host: str, step: int, task: int, job: int, measurement: INFLUXDB_MEASUREMENT
+) -> list[InfluxDBPointDict]:
+    """
+    Fetch data from InfluxDB for a given host, step and task.
+    """
+    query = f"""
+    SELECT * FROM {measurement} WHERE time > $time AND host = $host AND step = $step AND task = $task AND job = $job
+    """
+    with JobbergateApiError.handle_errors("Failed to fetch data from InfluxDB", do_except=log_error):
+        assert influxdb_client is not None  # mypy assertion
+        params = dict(time=time, host=host, step=str(step), task=str(task), job=str(job))
+        logger.debug(f"Querying InfluxDB with: {query=}, {params=}")
+        result = influxdb_client.query(query, bind_params=params, epoch="us")
+        logger.debug("Successfully fetched data from InfluxDB")
+        return [
+            InfluxDBPointDict(
+                time=point["time"],
+                host=point["host"],
+                job=point["job"],
+                step=point["step"],
+                task=point["task"],
+                value=point["value"],
+                measurement=measurement,
+            )
+            for point in result.get_points()
+        ]
+
+
+def fetch_influx_measurements() -> list[InfluxDBMeasurementDict]:
+    """
+    Fetch measurements from InfluxDB.
+    """
+    with JobbergateApiError.handle_errors("Failed to fetch measurements from InfluxDB", do_except=log_error):
+        logger.debug("Fetching measurements from InfluxDB")
+        assert influxdb_client is not None
+        measurements: list[InfluxDBMeasurementDict] = influxdb_client.get_list_measurements()
+        logger.debug(f"Fetched measurements from InfluxDB: {measurements=}")
+        return measurements
+
+
+async def update_job_metrics(active_job_submittion: ActiveJobSubmission) -> None:
+    """Update job metrics for a job submission.
+
+    This function fetches the metrics from InfluxDB and sends to the API.
+    """
+    with JobbergateApiError.handle_errors(
+        f"Could not update job metrics for slurm job {active_job_submittion.slurm_job_id} via the API",
+        do_except=log_error,
+    ):
+        response = await jobbergate_api_client.get(
+            f"jobbergate/job-submissions/agent/metrics/{active_job_submittion.id}"
+        )
+        response.raise_for_status()
+        job_max_times = JobSubmissionMetricsMaxResponse(**response.json())
+
+        influx_measurements = fetch_influx_measurements()
+
+        tasks = (
+            fetch_influx_data(
+                job_max_time.max_time,
+                job_max_time.node_host,
+                job_max_time.step,
+                job_max_time.task,
+                active_job_submittion.slurm_job_id,
+                measurement["name"],
+            )
+            for job_max_time in job_max_times.max_times
+            for measurement in influx_measurements
+        )
+        results = await asyncio.gather(*list(tasks))
+        data_points = chain.from_iterable(results)
+        aggregated_data_points = aggregate_influx_measures(data_points)
+        packed_data = msgpack.packb(aggregated_data_points)
+
+        response = await jobbergate_api_client.put(
+            f"jobbergate/job-submissions/agent/metrics/{active_job_submittion.id}",
+            content=packed_data,
+            headers={"Content-Type": "application/octet-stream"},
+        )
+        response.raise_for_status()
+
+
 async def update_active_jobs() -> None:
     """
     Update slurm job state for active jobs.
@@ -83,6 +178,14 @@ async def update_active_jobs() -> None:
 
     skip = "skipping to next active job"
     for active_job_submission in active_job_submissions:
+        if SETTINGS.influx_integration_enabled:
+            logger.debug(f"Updating job metrics for job_submission {active_job_submission.id}")
+            try:
+                await update_job_metrics(active_job_submission)
+            except Exception:
+                logger.error("Update job metrics failed... skipping for job data update")
+                pass
+
         logger.debug(f"Fetching slurm job state of job_submission {active_job_submission.id}")
 
         try: