+++ jobbergate-agent/tests/jobbergate/test_update.py
@@ -1,5 +1,10 @@
 import json
+import random
+from datetime import datetime
+from typing import get_args
 from unittest import mock
+from collections.abc import Callable
+import contextlib
 
 import httpx
 import pytest
@@ -11,9 +16,41 @@ from jobbergate_agent.jobbergate.update import (
     fetch_job_data,
     update_active_jobs,
     update_job_data,
+    fetch_influx_data,
+    fetch_influx_measurements,
+    update_job_metrics,
 )
+from jobbergate_agent.jobbergate.constants import INFLUXDB_MEASUREMENT
 from jobbergate_agent.settings import SETTINGS
 from jobbergate_agent.utils.exception import JobbergateApiError
+
+
+@pytest.fixture()
+def job_max_times_response() -> Callable[[int, int, int, int], dict[str, int | list[dict[str, int | str]]]]:
+    """Generates a sample response for the endpoint
+    ``jobbergate/job-submissions/agent/metrics/<job submission id>``.
+    """
+
+    def _job_max_times_response(
+        job_submission_id: int, num_hosts: int, num_steps: int, num_tasks: int
+    ) -> dict[str, int | list[dict[str, int | str]]]:
+        current_time = int(datetime.now().timestamp())
+        return {
+            "job_submission_id": job_submission_id,
+            "max_times": [
+                {
+                    "max_time": current_time,
+                    "node_host": f"host_{host}",
+                    "step": step,
+                    "task": task,
+                }
+                for host in range(1, num_hosts + 1)
+                for step in range(1, num_steps + 1)
+                for task in range(1, num_tasks + 1)
+            ],
+        }
+
+    return _job_max_times_response
 
 
 @pytest.mark.asyncio
@@ -194,7 +231,10 @@ async def test_update_job_data__raises_JobbergateApiError_if_the_response_is_not
 
 @pytest.mark.asyncio
 @pytest.mark.usefixtures("mock_access_token")
-async def test_update_active_jobs(mocker):
+async def test_update_active_jobs(
+    mocker,
+    tweak_settings: Callable[..., contextlib._GeneratorContextManager],
+):
     """
     Test that the ``update_active_jobs()`` function can fetch active job submissions,
     retrieve the job data from slurm, and update the slurm job data on the submission via the API.
@@ -203,14 +243,23 @@ async def test_update_active_jobs(mocker):
     mocked_sbatch = mock.MagicMock()
     mocker.patch("jobbergate_agent.jobbergate.update.InfoHandler", return_value=mocked_sbatch)
 
+    mocked_influxdb_client = mock.MagicMock()
+    mocker.patch("jobbergate_agent.jobbergate.update.influxdb_client", return_value=mocked_influxdb_client)
+
+    active_job_submissions = [
+        ActiveJobSubmission(id=1, slurm_job_id=11),  # Will update
+        ActiveJobSubmission(id=2, slurm_job_id=22),  # fetch_job_data throws exception
+        ActiveJobSubmission(id=3, slurm_job_id=33),  # update_job_data throws exception
+    ]
+
     mocker.patch(
         "jobbergate_agent.jobbergate.update.fetch_active_submissions",
-        return_value=[
-            ActiveJobSubmission(id=1, slurm_job_id=11),  # Will update
-            ActiveJobSubmission(id=2, slurm_job_id=22),  # fetch_job_data throws exception
-            ActiveJobSubmission(id=3, slurm_job_id=33),  # update_job_data throws exception
-        ],
+        return_value=active_job_submissions,
     )
+
+    def _mocked_update_job_metrics(active_job_submission: ActiveJobSubmission):
+        if active_job_submission.slurm_job_id == 11:
+            raise Exception("CRASH!")
 
     def _mocked_fetch_job_data(slurm_job_id, *args, **kwargs):
         if slurm_job_id == 22:
@@ -234,19 +283,30 @@ async def test_update_active_jobs(mocker):
         if job_submission_id == 3:
             raise Exception("BANG!")
 
-    mock_fetch = mocker.patch("jobbergate_agent.jobbergate.update.fetch_job_data", side_effect=_mocked_fetch_job_data)
-    mock_update = mocker.patch(
+    mock_update_job_metrics = mocker.patch(
+        "jobbergate_agent.jobbergate.update.update_job_metrics", side_effect=_mocked_update_job_metrics
+    )
+    mock_fetch_job_data = mocker.patch(
+        "jobbergate_agent.jobbergate.update.fetch_job_data", side_effect=_mocked_fetch_job_data
+    )
+    mock_update_job_data = mocker.patch(
         "jobbergate_agent.jobbergate.update.update_job_data", side_effect=_mocked_update_job_data
     )
 
-    await update_active_jobs()
+    with tweak_settings(INFLUX_DSN="https://influxdb:8086"):
+        await update_active_jobs()
 
-    mock_fetch.assert_has_calls(
+    mock_update_job_metrics.assert_has_calls(
+        [mocker.call(active_job_submission) for active_job_submission in active_job_submissions]
+    )
+    assert mock_update_job_metrics.call_count == 3
+
+    mock_fetch_job_data.assert_has_calls(
         [mocker.call(11, mocked_sbatch), mocker.call(22, mocked_sbatch), mocker.call(33, mocked_sbatch)]
     )
-    assert mock_fetch.call_count == 3
+    assert mock_fetch_job_data.call_count == 3
 
-    mock_update.assert_has_calls(
+    mock_update_job_data.assert_has_calls(
         [
             mocker.call(
                 1,
@@ -268,4 +328,370 @@ async def test_update_active_jobs(mocker):
             ),
         ]
     )
-    assert mock_update.call_count == 2
+    assert mock_update_job_data.call_count == 2
+
+
+@pytest.mark.asyncio
+@mock.patch("jobbergate_agent.jobbergate.update.influxdb_client")
+async def test_fetch_influx_data__success(mocked_influxdb_client: mock.MagicMock):
+    """
+    Test that the ``fetch_influx_data()`` function can successfully retrieve
+    data from InfluxDB as a list of ``InfluxDBPointDict``.
+    """
+    time = random.randint(0, 1000)  # noqa: F811
+    host = "test-host"
+    step = random.randint(0, 1000)
+    task = random.randint(0, 1000)
+    job = random.randint(0, 1000)
+    measurement_value = random.uniform(1, 1000)
+    measurement = random.choice(get_args(INFLUXDB_MEASUREMENT))
+
+    mocked_influxdb_client.query.return_value.get_points.return_value = [
+        dict(
+            time=time,
+            host=host,
+            job=job,
+            step=step,
+            task=task,
+            value=measurement_value,
+        )
+    ]
+
+    result = await fetch_influx_data(
+        time=time,
+        host=host,
+        step=step,
+        task=task,
+        job=job,
+        measurement=measurement,
+    )
+
+    assert len(result) == 1
+    assert result[0]["time"] == time
+    assert result[0]["host"] == host
+    assert result[0]["job"] == job
+    assert result[0]["step"] == step
+    assert result[0]["task"] == task
+    assert result[0]["value"] == measurement_value
+    assert result[0]["measurement"] == measurement
+
+
+@pytest.mark.asyncio
+@mock.patch("jobbergate_agent.jobbergate.update.influxdb_client")
+async def test_fetch_influx_data__raises_JobbergateApiError_if_query_fails(mocked_influxdb_client: mock.MagicMock):
+    """
+    Test that the ``fetch_influx_data()`` function will raise a JobbergateApiError
+    if the query to InfluxDB fails.
+    """
+    measurement = random.choice(get_args(INFLUXDB_MEASUREMENT))
+
+    mocked_influxdb_client.query = mock.Mock(side_effect=Exception("BOOM!"))
+
+    time = random.randint(0, 1000)  # noqa: F811
+    host = "test-host"
+    step = random.randint(0, 1000)
+    task = random.randint(0, 1000)
+    job = random.randint(0, 1000)
+
+    query = f"""
+    SELECT * FROM {measurement} WHERE time > $time AND host = $host AND step = $step AND task = $task AND job = $job
+    """
+    params = dict(time=time, host=host, step=str(step), task=str(task), job=str(job))
+
+    with pytest.raises(JobbergateApiError, match="Failed to fetch data from InfluxDB"):
+        await fetch_influx_data(
+            time=time,
+            host=host,
+            step=step,
+            task=task,
+            job=job,
+            measurement=measurement,
+        )
+
+    mocked_influxdb_client.query.assert_called_once_with(query, bind_params=params, epoch="us")
+
+
+@pytest.mark.asyncio
+async def test_fetch_influx_data__raises_JobbergateApiError_if_influxdb_client_is_None():
+    """
+    Test that the ``fetch_influx_data()`` function will raise a JobbergateApiError
+    if the influxdb_client is None.
+    """
+    measurement = random.choice(get_args(INFLUXDB_MEASUREMENT))
+    with mock.patch("jobbergate_agent.jobbergate.update.influxdb_client", None):
+        with pytest.raises(JobbergateApiError, match="Failed to fetch data from InfluxDB"):
+            await fetch_influx_data(
+                time=random.randint(0, 1000),
+                host="test-host",
+                step=random.randint(0, 1000),
+                task=random.randint(0, 1000),
+                job=random.randint(0, 1000),
+                measurement=measurement,
+            )
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize(
+    "measurements",
+    [
+        [{"name": "measurement1"}, {"name": "measurement2"}],
+        [{"name": "measurement1"}],
+        [],
+    ],
+)
+@mock.patch("jobbergate_agent.jobbergate.update.influxdb_client")
+async def test_fetch_influx_measurements__success(
+    mocked_influxdb_client: mock.MagicMock, measurements: list[dict[str, str]]
+):
+    """
+    Test that the ``fetch_influx_measurements()`` function can successfully retrieve
+    measurements from InfluxDB.
+    """
+    mocked_influxdb_client.get_list_measurements.return_value = measurements
+
+    result = fetch_influx_measurements()
+
+    assert result == measurements
+
+
+@pytest.mark.asyncio
+async def test_fetch_influx_measurements__raises_JobbergateApiError_if_influxdb_client_is_None():
+    """
+    Test that the ``fetch_influx_measurements()`` function will raise a JobbergateApiError
+    if the influxdb_client is None.
+    """
+    with mock.patch("jobbergate_agent.jobbergate.update.influxdb_client", None):
+        with pytest.raises(JobbergateApiError, match="Failed to fetch measurements from InfluxDB"):
+            fetch_influx_measurements()
+
+
+@pytest.mark.asyncio
+@mock.patch("jobbergate_agent.jobbergate.update.influxdb_client")
+async def test_fetch_influx_measurements__raises_JobbergateApiError_if_query_fails(
+    mocked_influxdb_client: mock.MagicMock,
+):
+    """
+    Test that the ``fetch_influx_measurements()`` function will raise a JobbergateApiError
+    if the query to InfluxDB fails.
+    """
+    mocked_influxdb_client.get_list_measurements.side_effect = Exception("BOOM!")
+
+    with pytest.raises(JobbergateApiError, match="Failed to fetch measurements from InfluxDB"):
+        fetch_influx_measurements()
+
+    mocked_influxdb_client.get_list_measurements.assert_called_once_with()
+
+
+@pytest.mark.asyncio
+@pytest.mark.usefixtures("mock_access_token")
+@pytest.mark.parametrize(
+    "job_submission_id, slurm_job_id",
+    [
+        (1, 22),
+        (2, 33),
+        (3, 11),
+    ],
+)
+async def test_update_job_metrics__error_getting_metrics_from_api(job_submission_id: int, slurm_job_id: int):
+    """
+    Test that the ``update_job_metrics()`` function will log an error if it fails
+    to get the job metrics from the API.
+    """
+    active_job_submission = ActiveJobSubmission(id=job_submission_id, slurm_job_id=slurm_job_id)
+
+    with respx.mock:
+        respx.get(f"{SETTINGS.BASE_API_URL}/jobbergate/job-submissions/agent/metrics/{job_submission_id}").mock(
+            return_value=httpx.Response(status_code=400)
+        )
+
+        with pytest.raises(
+            JobbergateApiError, match=f"Could not update job metrics for slurm job {slurm_job_id} via the API"
+        ):
+            await update_job_metrics(active_job_submission)
+
+
+@pytest.mark.asyncio
+@pytest.mark.usefixtures("mock_access_token")
+@pytest.mark.parametrize(
+    "job_submission_id, slurm_job_id, num_hosts, num_steps, num_tasks, measurements",
+    [
+        (1, 22, 5, 2, 7, [{"name": "measurement1"}, {"name": "measurement2"}]),
+        (2, 33, 1, 1, 1, [{"name": "measurement1"}]),
+        (3, 11, 3, 10, 4, [{"name": "measurement1"}, {"name": "measurement2"}, {"name": "measurement3"}]),
+    ],
+)
+@mock.patch("jobbergate_agent.jobbergate.update.fetch_influx_measurements")
+@mock.patch("jobbergate_agent.jobbergate.update.fetch_influx_data")
+@mock.patch("jobbergate_agent.jobbergate.update.aggregate_influx_measures")
+@mock.patch("jobbergate_agent.jobbergate.update.msgpack")
+@mock.patch("jobbergate_agent.jobbergate.update.chain")
+async def test_update_job_metrics__error_sending_metrics_to_api(
+    mocked_chain: mock.MagicMock,
+    mocked_msgpack: mock.MagicMock,
+    mocked_aggregate_influx_measures: mock.MagicMock,
+    mocked_fetch_influx_data: mock.MagicMock,
+    mocked_fetch_influx_measurements: mock.MagicMock,
+    job_submission_id: int,
+    slurm_job_id: int,
+    num_hosts: int,
+    num_steps: int,
+    num_tasks: int,
+    measurements: list[dict[str, str]],
+    job_max_times_response: Callable[[int, int, int, int], dict[str, int | list[dict[str, int | str]]]],
+):
+    """
+    Test that the ``update_job_metrics()`` function will log an error if it fails
+    to send the job metrics to the API.
+    """
+    active_job_submission = ActiveJobSubmission(id=job_submission_id, slurm_job_id=slurm_job_id)
+    job_max_times = job_max_times_response(job_submission_id, num_hosts, num_steps, num_tasks)
+
+    dummy_data_point = {
+        "time": 1,
+        "host": "host_1",
+        "job": "1",
+        "step": "1",
+        "task": "1",
+        "value": 1.0,
+        "measurement": "measurement1",
+    }
+    dummy_data_points = [dummy_data_point] * len(measurements) * len(job_max_times["max_times"])
+    iter_dummy_data_points = iter(dummy_data_points)
+
+    mocked_fetch_influx_measurements.return_value = measurements
+    mocked_fetch_influx_data.return_value = dummy_data_points
+    # doesn't return the real aggregated data due to test complexity
+    mocked_chain.from_iterable.return_value = iter_dummy_data_points
+    mocked_aggregate_influx_measures.return_value = "super-dummy-aggregated-data"
+    mocked_msgpack.packb.return_value = b"dummy-msgpack-data"
+
+    with respx.mock:
+        respx.get(f"{SETTINGS.BASE_API_URL}/jobbergate/job-submissions/agent/metrics/{job_submission_id}").mock(
+            return_value=httpx.Response(
+                status_code=200,
+                json=job_max_times,
+            )
+        )
+        respx.put(
+            f"{SETTINGS.BASE_API_URL}/jobbergate/job-submissions/agent/metrics/{job_submission_id}",
+            content=b"dummy-msgpack-data",
+            headers={"Content-Type": "application/octet-stream"},
+        ).mock(return_value=httpx.Response(status_code=400))
+
+        with pytest.raises(
+            JobbergateApiError, match=f"Could not update job metrics for slurm job {slurm_job_id} via the API"
+        ):
+            await update_job_metrics(active_job_submission)
+
+    mocked_fetch_influx_measurements.assert_called_once_with()
+    mocked_fetch_influx_data.assert_has_calls(
+        [
+            mock.call(
+                job_max_time["max_time"],
+                job_max_time["node_host"],
+                job_max_time["step"],
+                job_max_time["task"],
+                slurm_job_id,
+                measurement["name"],
+            )
+            for job_max_time in job_max_times["max_times"]
+            for measurement in measurements
+        ]
+    )
+    mocked_chain.from_iterable.assert_called_once_with(
+        [dummy_data_points] * len(measurements) * len(job_max_times["max_times"])
+    )
+    mocked_aggregate_influx_measures.assert_called_once_with(iter_dummy_data_points)
+    mocked_msgpack.packb.assert_called_once_with("super-dummy-aggregated-data")
+
+
+@pytest.mark.asyncio
+@pytest.mark.usefixtures("mock_access_token")
+@pytest.mark.parametrize(
+    "job_submission_id, slurm_job_id, num_hosts, num_steps, num_tasks, measurements",
+    [
+        (1, 22, 5, 2, 7, [{"name": "measurement1"}, {"name": "measurement2"}]),
+        (2, 33, 1, 1, 1, [{"name": "measurement1"}]),
+        (3, 11, 3, 10, 4, [{"name": "measurement1"}, {"name": "measurement2"}, {"name": "measurement3"}]),
+    ],
+)
+@mock.patch("jobbergate_agent.jobbergate.update.fetch_influx_measurements")
+@mock.patch("jobbergate_agent.jobbergate.update.fetch_influx_data")
+@mock.patch("jobbergate_agent.jobbergate.update.aggregate_influx_measures")
+@mock.patch("jobbergate_agent.jobbergate.update.msgpack")
+@mock.patch("jobbergate_agent.jobbergate.update.chain")
+async def test_update_job_metrics__success(
+    mocked_chain: mock.MagicMock,
+    mocked_msgpack: mock.MagicMock,
+    mocked_aggregate_influx_measures: mock.MagicMock,
+    mocked_fetch_influx_data: mock.MagicMock,
+    mocked_fetch_influx_measurements: mock.MagicMock,
+    job_submission_id: int,
+    slurm_job_id: int,
+    num_hosts: int,
+    num_steps: int,
+    num_tasks: int,
+    measurements: list[dict[str, str]],
+    job_max_times_response: Callable[[int, int, int, int], dict[str, int | list[dict[str, int | str]]]],
+):
+    """
+    Test that the ``update_job_metrics()`` function will log an error if it fails
+    to send the job metrics to the API.
+    """
+    active_job_submission = ActiveJobSubmission(id=job_submission_id, slurm_job_id=slurm_job_id)
+    job_max_times = job_max_times_response(job_submission_id, num_hosts, num_steps, num_tasks)
+
+    dummy_data_point = {
+        "time": 1,
+        "host": "host_1",
+        "job": "1",
+        "step": "1",
+        "task": "1",
+        "value": 1.0,
+        "measurement": "measurement1",
+    }
+    dummy_data_points = [dummy_data_point] * len(measurements) * len(job_max_times["max_times"])
+    iter_dummy_data_points = iter(dummy_data_points)
+
+    mocked_fetch_influx_measurements.return_value = measurements
+    mocked_fetch_influx_data.return_value = dummy_data_points
+    # doesn't return the real aggregated data due to test complexity
+    mocked_chain.from_iterable.return_value = iter_dummy_data_points
+    mocked_aggregate_influx_measures.return_value = "super-dummy-aggregated-data"
+    mocked_msgpack.packb.return_value = b"dummy-msgpack-data"
+
+    with respx.mock:
+        respx.get(f"{SETTINGS.BASE_API_URL}/jobbergate/job-submissions/agent/metrics/{job_submission_id}").mock(
+            return_value=httpx.Response(
+                status_code=200,
+                json=job_max_times,
+            )
+        )
+        respx.put(
+            f"{SETTINGS.BASE_API_URL}/jobbergate/job-submissions/agent/metrics/{job_submission_id}",
+            content=b"dummy-msgpack-data",
+            headers={"Content-Type": "application/octet-stream"},
+        ).mock(return_value=httpx.Response(status_code=200))
+
+        await update_job_metrics(active_job_submission)
+
+    mocked_fetch_influx_measurements.assert_called_once_with()
+    mocked_fetch_influx_data.assert_has_calls(
+        [
+            mock.call(
+                job_max_time["max_time"],
+                job_max_time["node_host"],
+                job_max_time["step"],
+                job_max_time["task"],
+                slurm_job_id,
+                measurement["name"],
+            )
+            for job_max_time in job_max_times["max_times"]
+            for measurement in measurements
+        ]
+    )
+    mocked_chain.from_iterable.assert_called_once_with(
+        [dummy_data_points] * len(measurements) * len(job_max_times["max_times"])
+    )
+    mocked_aggregate_influx_measures.assert_called_once_with(iter_dummy_data_points)
+    mocked_msgpack.packb.assert_called_once_with("super-dummy-aggregated-data")